Add - Copy tar to container, create acc list
When openstack container is added to mc host list as eukupload, we now
copy or archives to it. Initially, it will be to folder batch1. Maybe
that will change in future. I also keep track which accessions were
processed in which set be copying separate file with just accessions to
same folder as archives. Rule to create such file is also added here.