Add rule to download protein profiles, Snakemake upd v5.9.1 from 5.7.0
Updated Snakemake to last available version at the moment. Updates have
better support of HPC and other stuff that may be useful.

Rule to download protein profiles for MetaEuk. Before, I manually
downloaded once these profiles from Elli's site to /mnt folder.
But now it is important to be done automatically for first start
of any new instance from bibigrid cluster, as it can't be stored in an
instance image, like other software and configuration, because it is
~160Gb large and won't fit on 20 Gb disk, that image is taken of.