Make megahit rule upload it's results to remote server
Magahit rule in addition to what it did before is also creating an
archive containing only the assembled contigs and log file, and uploads
it to remote server's "megahit" folder.

In this way I want to create a backup for all assemblies on server to
run future MetaEuk or EukRep programs with different parameters without
running again megahit, or downloading whole previous pipeline with
previous result of EukRep and Megahit to override them. Same approach
awaits outputs of Eukrep, MultiQC and MEtaeuk ofcource. In this way the
backup is modular.

This was more dificult than expected. One serious issue took a lot of
time to solve. Each megahit when finished was deleting the local copy of
eukcontainer/megahit folders, when uploaded it's own file to server,
whereas other megahits were still running and also needed that folder.
To keep folder from deleting, I ask each megahit to put an empty file to
this folder, so no megahit that finished earlier will delete the
folder, because it's not empty.

I create this tar with pigz, so I had to add it to megahit environment.