Make Snakefile more clear, readible, tidy.
Clearly separated Snakefile in three parts:
Config, Input functions, and Workflow.
This was required because of growing complexity of snakefile.
And also because of input functions. I think it's a good thing to
clearly define all of them in separate dedicated part of snakefile.
Especially because there will be more.

Added explicit comments to many things to make it more clear.

Formatted code more nicely.