Remove megahit_download fork
I just realised that Snakemake keeps the local copy of remote files if
some later rules depend on them. That means it doesn't try to download
back and forth the same files, so I don't really need a fork. I do need
however to play around with compressing and decompressing, but this is
minor issue in the scope of whole pipeline and doesn't matter so much.

I am really happy Snakemake keeps track of the remote files that will
be needed by later rules and keeps them. This means I don't need the
megahit_download rule and input functions and I will delete them.