Split megahit rule into megahit and megahit_upload
This will allow to only upload the alreadly calculated local megahits to
remote server when running Snakemake on old directories. Otherwise, to
have a remote copy of megahit, the whole megahit job would be run even
if a local copy was already present.

Also changed QUAST_raw_contigs rule to operate with parameters. It is
nicer now.