Add save sample list to remote, reformat MetaEuk rule
Sample list is a simple txt file containing list of accessions used for
a MetaEuk run. It is stored within MetaEuk.tar.gz results archive on
remote server, as well as just a separate txt file near the archive to
have an easy access if you just need to read the accessions.

These sample lists will be used by a scheduling script. The script will
read all such files in Metaeuk results remote folder and decide which
submitted samples have already been processed and should be ommited in
pipeline, and which should be given to Snakemake parallel jobs.

Rearranged the namings in MetaEuk rule. Tarred it in the same way as
other tars to avoid having set_name folder inside the tar.