Specify files for MultiQC and improve its runtime
Before MultiQC was set to search whole set_name directory for report
files. With all temp files, many of which are big, it was ~1700 files for
5 samples and it took quite some time (up to 10 min) for MultiQC to search
through all the files. Specifying what files and folders should it look
at shortened the runtime to 10-20 seconds.

Also changed the file name for contigs filtered by size from sizefilt to
filtsize. In this way it is alphabetically sorted in a way that is also
logical to compare in myltiqc report.