Update set.yaml preparation script
The script is better now.

NOTE:    It counts for the set_*_samples.txt files at
remote container, not on processed archives. So in case of error you
will get set accessions uploaded, but no archive uploaded. This set
will seem like processed for script, when the set was in fact not processed and
uploaded to the end. These cases have to be followed and rerun manually.

Additionally added 4581 sample list without quotes - they break the set
preparation process.