Make QUASTs avoid deleting results and log on error
log file is stored in the results directory of quast, that is deleted on
any error by snakemake. So there is no way to look at what exactly
caused the error.

If the results directory is specified via params and the only output
is the tar file to S3 remote server, then the results folder is not
deleted on error and you can look at the log file after error.