Avoid QUAST errors
Many sets failed before because some small sample at some point had no
contigs in filtered contigs file. If empty files are input for QUAST, it
throws an error and breaks all the pipeline. But in some cases, as for
small samples, it is natural not to have any big contigs or any
eukaryotic contigs, so this should be passed safely by the pipeline.

Here I implemented avoiding such errors. First quast checks if input
file is empty or not. If it is empty, quast creates folder with log only
and this is saved to remote container without throwing an error.
Following MultiQC rule just omits QUAST result of such sample.

NOTE: I skip only "empty input" file errors. All other errors will stil
break the pipeline, and that is intentional. I don't want to miss QUAST
results if they were canceled not due to empty file error but to smth
else that I will have to see and fix. Otherwise I would be unaware of
new problems.