Make set_asmples.txt be uploaded only in very end
In this way you could tell from files on remoe container which sets were
successfully processed to the very end. Also, the change date of the
set_samples.txt must be the earliest compared to all others outut of the
set. If something of above is wrong - probably set (re-)run ended
unsuccessfully.

ATTENTION: now the set_preparation script will not see the set lists of
unprocessed files. So theoretically, if the last set (with the biggest
number) did not finesh successfully, the preparation script will think
that there are no set using this number (eg last - 99), and will create
sets from 99 for new accessions input. This will cause overwriting any
unfinished set results for set 99 and missing out the previous
accesssions for this number.