Make Megahit --continue if previous run was interupted
Before, if any error hapened during megahit run in its rule, entire
output was deleted. Megahit is the most resource-consuming job of
pipeline, so it is a big waste of resources.

Luckily, Megahit has an option to --continue from the last
checkpoint, if it is run in same directory. Previously the directory was
a part of output and was deleted upon an error. Now it is a parameter
and only the contigs are the output.

Megahit rule automatically decides how to launch megahit depending on
the presence of the output folder and the checkpoint file "done".