Implement Megahit_parallel_mode as config option
Big samples keep failing the pipeline. I decided to run them with "fix"
snakemake script that has maximum resources allowed to the megahit.
Thus, it is with max memory and not parallel.

Additionally, I wanted to keep the parallel option as default to all
other sets. So, I decided to put megahit_parallel execution as an
option. In my config now it is a default True. Fix script overwrites it
to be False.