Make snakemake add workflow stamps to master node
In case if mailing doesn't work, there was now way to have a general
overview of all started/finished/errored Snakemakes. In this case would
have to look which sets are missing in final output folder of remote
server, and after thet go finding which of the nodes executed given set
to see what was the problem.

Now, in addition to mailing, snakemake pipeline puts stamps to the
master node folder ~/snakemake_stamps when it starts, ends, or has an
error, similarly to the mailing system. The stamp name contains the set
number and the worker node name. Date of the stamp modification matches
the time when the representatinve pipeline started, ended, or errored,
correspondingly.