Make MetaEuk output aa sequence predictions
Explanation:
Before for metaeuk easypredict step I missed the parameter --protein, so
all the final metaeuk output prediction fasta files contained nucleotide
sequence for predicted proteins rather that aminoacid sequnce. The whole
EukBook project for 4,573 samples was run like that.

This is not a big deal for prediction itself, it is possible to simple
convert the nucleotide sequences into aa. But the taxonomy stage was
affected. MetaEuk taxonomic prediction on nucleotide sequence works
differently from aa sequence, because in the first case it performs
6-frame translation and predicts the taxonomy by all putative short
proteins that could arise from such translation, not only the main one.

So, taxonomic prediction for first run of EukBook project was not correct
and should be redone.

Main hange:
1. This commit adds the --protein flag to metaeuk easy-predict to ensure
the output fasta file is in aa format and taxonomy is done properly.

Small changes:
2. To not overwrite the previous prediction and taxonomy results, I save
the new results into a different folder of metaeuk - metaeul_alpha_aa.
This is done by changing the "metaeuk version", although the actual
version of MetaEuk is the same.

3. Changed the output of metaeuk easy-predict to have a set_n_ in the
filename as the utput of metaeuk taxtocontig.

4. Changed time command to /usr/bin/time for both metaeuk commands and
the Snakemake wrapper itself. So now more resource usage statistics is
collected, moreover, the overall pipeline statistics is collected and
peak memory usage per pipeline, that is great. That could have helped to
monitor the Megahit crashes before.